Daily Learning should not trigger addHebrewDatesForEvents.
Hebrew dates only showing when it's a Holiday of some sort (Rosh Chodesh, Yamim Tovim, etc -
just to not have the overload of dates every weekday) is helpful.

This fix changes the setting "Show Hebrew date for dates with some event" to suppress these
events if the only event on a given day is dailyLearning.